move issue tracker interface definition to seperate file to prepare for better integration
