+++ .gitignore
@@ -176,3 +176,4 @@ cython_debug/
 
 # IDE settings
 .vscode/
+test-results/
